+++ examples/EventSourcedEntity/run-me.php
@@ -1,0 +1,9 @@
+<?php
+
+require_once('../../var/vendor/autoload.php');
+require_once('User.php');
+require_once('UserCreatedWithUsername.php');
+
+$user = User::createWithUsername("David");
+
+echo "Hello, {$user->username()}!" . PHP_EOL;