+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.1.2
+ * Added an example entity using EventSourcedEntityMixin
+
 ## 0.1.1
-Initial release (Need to begin somewhere!)
+ * Initial release (Need to begin somewhere!)